+++ server/src/middleware/fileHandler/modelProfileHandler.ts
@@ -3,6 +3,7 @@ import * as fs from "fs";
 const fsPromises = fs.promises;
 import log from "../../utils/logging/logger";
 import modelService from "../../service/model/index.service";
+import { DEFAULT_MODEL_PROFILE_IMAGE } from "../../utils/constants";
 
 export const uploadModelProfile = async (
   req: Request,
@@ -57,8 +58,12 @@ export const deleteModelProfile = async (req: Request, res: Response, next: Next
   try{
     const rawModel = await modelService.findOneById(req.params.model_id);
     const model = await  rawModel.dataValues;
+    
     for(let i = 1; i < 7; i++){
       if(model[`profile_img_${i}`] !== null){
+        if(model[`profile_img_${i}`] === DEFAULT_MODEL_PROFILE_IMAGE) {
+          continue;
+        }
        try{
          await fsPromises.unlink(model[`profile_img_${i}`]);
        }catch(err){
